+++ config.example.php
@@ -21,14 +21,3 @@ define('EDITH_URI', '');
 
 // Extension of files living in the data directory
 # define('EDITH_DATA_EXTENSION', '.txt');
-
-/*
- * To handle concurrent access, you can enable Google Mobwrite by using the lines below.
- *  Read more about Google Mobwrite: https://code.google.com/archive/p/google-mobwrite/
- */
-
-// Choose a unique identifier for your application so that Mobwrite can identify all your pages.
-# define('MOBWRITE_KEY', 'edith');
-
-// Mobwrite URL
-# define('MOBWRITE_URI', 'https://edith-mobwrite.appspot.com');