+++ index.php
@@ -33,12 +33,6 @@ if (!defined('URI_REGEX'))
     '#^/?([^/]+?)\.?('.implode('|', EDITH_REPRESENTATIONS).')?$#'
   );
 
-if (!defined('MOBWRITE_KEY'))
-  define('MOBWRITE_KEY', 'edith');
-
-if (!defined('MOBWRITE_URI'))
-  define('MOBWRITE_URI', null);
-
 if (file_exists(preg_replace('#^\/#', '', $_SERVER["REQUEST_URI"])))
   return false;